+++ packages/filterpane/src/components/CheckBoxFilter/CheckBoxOption.tsx
@@ -1,32 +1,7 @@
 import { ReactNode } from 'react';
-import { createStyles, makeStyles } from '@equinor/fusion-react-styles';
 import CheckBox from '@equinor/fusion-react-checkbox';
 
-const useStyles = makeStyles(
-  () =>
-    createStyles({
-      FilterOption: {
-        display: 'flex',
-        alignItems: 'center',
-        cursor: 'pointer',
-      },
-      FilterOptionLabel: {
-        flex: 1,
-        overflow: 'hidden',
-        height: '100%',
-        cursor: 'pointer',
-        alignItems: 'center',
-        textOverflow: 'ellipsis',
-      },
-      FilterOptionCount: {
-        wordBreak: 'keep-all',
-        display: 'flex',
-        padding: '8px',
-        justifyContent: 'flex-end',
-      },
-    }),
-  { name: 'fusion-filterpane-checkboxFilterOption' }
-);
+import { useStyles } from './CheckBoxOption.style';
 
 type CheckboxOptionProps = {
   filterKey: string;
@@ -39,7 +14,7 @@ type CheckboxOptionProps = {
   singleSelect?: boolean;
 };
 
-const CheckboxOption = ({
+export const CheckboxOption = ({
   filterKey,
   onSelectionChange,
   selected,