+++ packages/radio/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.6](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-radio@0.3.5...@equinor/fusion-react-radio@0.3.6) (2021-10-05)
+
+**Note:** Version bump only for package @equinor/fusion-react-radio
+
+
+
+
+
+## 0.3.5 (2021-09-30)
+
+**Note:** Version bump only for package @equinor/fusion-react-radio
+
+
+
+
+
 ## [0.3.4](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-radio@0.3.3...@equinor/fusion-react-radio@0.3.4) (2021-09-27)